Put Thread in ctr_main instead.
